requirements.txt
# Core scraping dependencies
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.3

# Configuration management
python-dotenv>=1.0.0

# System monitoring & process management
psutil>=5.9.0

# Web dashboard
Flask>=3.0.0
Flask-Login>=0.6.3
Werkzeug>=3.0.0

# Task scheduling
APScheduler>=3.10.0

# Data handling
pandas>=2.1.0

# Optional but recommended for production
# prometheus-client>=0.19.0  # Metrics export
# sentry-sdk>=1.39.0         # Error tracking
